Add WIL summary and purpose to README, add Code of Conduct doc
If you're wondering what WIL is and why you'd want it,
now the README will give you some hints!

The Code of Conduct doc simply repeats what is already in the README,
which in turn refers to the Microsoft Open Source Code of Conduct
on https://opensource.microsoft.com . However, GitHub only recognizes
separate Code of Conduct doc files and not Code of Conduct sections
in a README, so this PR adds one.