Fix noreturn compile error for clang
Many of the functions called from the RETURN_IF_FAILED macro used noreturn annotation since in the failfast and exception case it isn't expected for them to return. In the case that a failfast or exception wasn't expected the functions would still retrun and had SAL annotation to express this. Clang doesn't pick up on the SAL resulting in a warning being generated. This change provides template overloads based on the return type so that in all cases the functions are properly annotated as noreturn or not. Additionally, the suppress action flag was factored into a template parameter for ReportFailure as that overrides no return behavior and setting it at runtime for a failfast or exception return would result in UB. Tested on VS 2017 and Clang 9.0.0