Fix noreturn compile error for clang (#80)
* Fix noreturn compile error for clang

Many of the functions called from the RETURN_IF_FAILED macro used noreturn annotation since in the failfast and exception case it isn't expected for them to return. In the case that a failfast or exception wasn't expected the functions would still retrun and had SAL annotation to express this. Clang doesn't pick up on the SAL resulting in a warning being generated. This change provides template overloads based on the return type so that in all cases the functions are properly annotated as noreturn or not. Additionally, the suppress action flag was factored into a template parameter for ReportFailure as that overrides no return behavior and setting it at runtime for a failfast or exception return would result in UB. Tested on VS 2017 and Clang 9.0.0

* Remove invalid SAL

Remove SAL that was invalidated due to making FailureType a template parameter. Also fix a build error caused by a non-updated instance of a function.

* Update cppwinrt.h

Fix compile error

* Retain existing interfaces for ReportFailure and ReportFailure_Hr

Some code takes a direct dependency on ReportFailure and ReportFailure_Hr outside of the failure macros. Need to retain those signatures so that the dependent code doesn't break and just have them defer to the templated implementation.

* Update result_macros.h

Fix compile error.

* Ensure that exception based macros are properly marked no-return

Not all of the exception based functions called from THROW_IF_FAILED were properly annotated with no return so ensure that they are on par with the functions called from RETURN_IF_FAILED. Also suppress C4127 around ReportFailure since if constexpr is only valid for newer standards.