PVS-studio static analyzer warnings fixes (#161)
* Comment ownerless token after preprocessor directive

* Add comment to empty catch everything handler

Done by analogy with other places were we fall through from empty catch everything handler.

* Fix UB when format string doesn't correspond to actual argument type

Both ::GetCurrentProcessId() and size are of DWORD aka unsigned long type,
but we have tried to format them as plain signed ints. That can lead to UB
in case its values > INT_MAX.

* Default initialize uninitialized thread id

There are not so many options here - either leave threadId uninitialized, which
results in UB on access to it, or set it to some very unlikely to be reached
value. Decided to go the second way.