wrl_module_reference and winrt_module_reference to prevent DLL from unloading
`wrl_module_reference` and `winrt_module_reference` hold a reference
to the host WRL or C++/WinRT module to prevent it from being unloaded.
Normally, the reference is held implicitly because you are a member
function of a DLL-hosted COM object, or because you are a coroutine
that returns IAsyncAction (or similar), or because you retain a strong
reference to some DLL-hosted object. But if those do not apply to you
then you will need to hold a reference explicitly.

Examples in comments.

WARNING: If `wrl_module_reference` is used from a module that does not
use WRL factories, it will have no effect. This is expected if used from
an executable module, but will be surprising if used from a DLL.