Avoid ODR violation if /await and non-/await code is linked together
MSVC C++17 pre-standard coroutines are opt-in, but it is legal
to combine code that uses them with code that doesn't.
(This is unlike MSVC C++17 pre-standard coroutines and MSVC C++20
coroutines, which are not compatible and cannot be combined.)

Therefore, we cannot define our own dummy `coroutine_handle`
when coroutine support is missing, because another translation unit
may compile with coroutine support enabled, and now we have an ODR
violation.

Instead, `#ifdef` out everything (except the traits class) if
coroutines are not supported. We leave the traits class so that
the "light-up" code can just specialize the traits class without
having to check whether coroutines are enabled.