wil::resume_foreground moved to cppwinrt_helpers.h to avoid forward-declaration collisions (#214)
* wil::resume_foreground moved to cppwinrt_helpers.h to avoid forward-declaration collisions

The forward declaration of the ::winrt::Microsoft namespace created
collisions with the ::Microsoft::WRL namespace when consumed from
inside the ::winrt namespace, because "Microsoft" resolves to
::winrt::Microsoft, but people who aren't using `winrt/Microsoft.*.h`
are expecting it to resolve to ::Microsoft.

We follow the `resource.h` pattern, where you include the headers
for the namespaces you care about, and then you include `cppwinrt_helpers.h`
to activate the helpers for those namespaces. You can include
`cppwinrt_headers.h` multiple times, and it will turn on the features
related to the namespaces you included since the last time you included
`cppwinrt_headers.h`.

This means that `cppwinrt_headers.h` does not introduce any namespaces.
It only takes advantage of pre-existing ones. If there is a namespace
name collision, it was introduced by the caller.

* Make it work on compilers with no coroutine support at all

* wil\cppwinrt_helpers.h no longe dependent upon wil\cppwinrt.h

This minimizes dependencies for people who would rather not
ingest the entire WIL error-reporting infrastructure.

* Avoid ODR violation if /await and non-/await code is linked together

MSVC C++17 pre-standard coroutines are opt-in, but it is legal
to combine code that uses them with code that doesn't.
(This is unlike MSVC C++17 pre-standard coroutines and MSVC C++20
coroutines, which are not compatible and cannot be combined.)

Therefore, we cannot define our own dummy `coroutine_handle`
when coroutine support is missing, because another translation unit
may compile with coroutine support enabled, and now we have an ODR
violation.

Instead, `#ifdef` out everything (except the traits class) if
coroutines are not supported. We leave the traits class so that
the "light-up" code can just specialize the traits class without
having to check whether coroutines are enabled.

Co-authored-by: Duncan Horn <40036384+dunhor@users.noreply.github.com>