co_await resume_any_apartment(async) resumes in any apartment
This lets you override the default "return to the original
apartment" behavior of `co_await`'ing an `IAsyncInfo`.

This is the C++/WinRT counterpart to C# `ConfigureAwait(false)`.

This is basically a copy of the C++/WinRT awaiter, but with
the one line of apartment-switching code deleted. Unfortunately,
it's a good amount of code that has to be duplicated.

The code is made a bit more complicated by the fact that it
has to deal with two sets of client options:

* C++/WinRT version 230207 introduced a breaking change in
  how custom awaiters support cancellation propagation, so
  we generate different cancellation support depending on which
  C++/WinRT version you are using.
* Avoiding a memory corruption bug in MSVC pre-standardization
  coroutines. This is fixed in VS version 16.11, but pre-16.11 and
  post-16.11 pre-standardization coroutines are ABI compatible, so
  changing behavior based on the VS version could lead to ODR violations.
  Instead, we key off the ABI breaking change in MSVC standardized
  coroutines in C++20, since the ABI break prevents you from linking
  pre-standardized coroutines with standardized coroutines. (This is
  the same solution that C++/WinRT itself uses.)

Fixes #274