Constrain the generator implementation to `IIterator<T>` and add `wil::make_iterable_from_iterator` to create an `IIterable<T>` helper object
C++ coroutines can only be evaluated once, which doesn't fulfill
`IIterable<T>`'s requirement that it can return multiple independent iterators.
Thus the generator implementation is constrained to `IIterator<T>`, which may
only be evaluated once, and a helper method is added that returns an
`IIterable<T>` which creates a new generator instance every time an iterator
is requested.