Fix malloc spy conflict with CppWinRTAuthoringTests::NotifyPropertyChanged (#368)
* Fix malloc spy conflict with CppWinRTAuthoringTests::NotifyPropertyChanged

The CppWinRTAuthoringTests::NotifyPropertyChanged test
did not clean up everything before it exited. This resulted
in COM remaining initialized at process termination, and
that in turn resulted in adverse interactions with the malloc
spy test, since the "emergency cleanup" calls CoTaskMemFree,
and since COM was put into "possible malloc spy" mode, it
tries to take a lock to protect the potential malloc spy.
But that lock no longer exists, since COM has shut down,
and things spiral downhill.

The NotifyPropertyChanged test tries to clean up the
WindowsXamlManager, but it didn't realize that Close()
is an asynchronous call that queues work back to the
current thread's DispatcherQueue for final cleanup.
The old test just returned immediately and never pumped
messages any more, which resulted in the cleanup never
happening.

To ensure we clean up the WindowsXamlManager, we create
our own DispatcherQueue for the XAML work. Since it's
our DispatcherQueue, we can call ShutdownQueueAsync
to clean it up. ShutdownQueueAsync completes when the
queue has shut down, but before the thread uninitializes COM.
We wait for the DispatcherQueue thread to exit, so that we
know for sure that it has definitely uninitialized COM.
Only then do we allow the test to complete.

This cleans up the leaked COM initialization and avoids
a condition crash during "emergency cleanup".

* Need to initialize as MTA to avoid sync wait on STA

* Don't wait indefinitely for DQ to shut down

Let the WFSO do the wait for both the shutdown and the thread exit.

Also check the TryEnqueue just for extra safety.

Now that we fixed test conflict, we can let the XAML test coexist with the mallocspy test.

---------

Co-authored-by: Duncan Horn <40036384+dunhor@users.noreply.github.com>