register_com_server: Fix incorrect CoRegisterClassObject rclsid argument
The implementation is currently calling CoRegisterClassObject with the default interface ID GUID as rclsid argument. It should instead be passing the COM class ID GUID (ClassID) as rclsid argument.

This commit will trigger an "error C2787: '<ClassName>': no GUID has been associated with this object" in  classes passed as template argument to register_com_server that lack a UUID.